requirements-dev.txt
codecov==2.0.9
coverage==4.4.2
flake8==3.5.0
mccabe==0.6.1
mock==2.0.0
pbr==3.1.1
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.2.5
pytest-flask==0.10.0
pytest-mock==1.6.3
six==1.11.0

-r requirements.txt

